bugfixed a rare issue with the game, and made the background darker
The bug was that sometimes on the first turn when you clicked on the
board an o would be placed instead of the x.   After this everything
worked fine, and the place where you put an o was still considered an x.
I think I fixed it by creating a board when the page loads, and simply
set it to be hidden until you hit a button to play against.